Add neovim config
Ngl, I didn't realize I accidently pushed a version of the nvim folder
that still had the LazyGit git repo, so it was pushed as a submodule. I
was getting annoyed at the inconsistency between my devices and was
trying to figure out the issue, finally looked at the Git repo on Gitea
and realized that lol. So I went ahead and cleaned config, and now we
should be good. Nvim config is in the best state it has even been in.