oh my god
I still need to figure out what to do with the .uuid font files (I
sucked with the gitignore ig) and btop since it saves state everytime
you toggle something, which is a little annoying. I'm debating if it
should even be here at all. I think all I used it for originally was the
theme lol.